requirements.txt
pytest
requests
selene==2.0.0rc9
allure-pytest==2.13.5